fix(authoring): land remaining Area 7 review nits
Clamp RepeatCount instead of ToInt32, treat DwellLimitMs as a double, reset a forgotten default report kind without materializing ReportKinds, and freeze formula/chrome copy strings in tests.

+++ src/HardwareTest.Authoring.Core/AuthoringMetricSettingCatalog.cs
@@ -47,7 +47,7 @@ public static class AuthoringMetricSettingCatalog
         ("*", "TsSeconds", new(AuthoringSettingKind.Double, Minimum: 0)),
         ("*", "FailWhenOutOfBand", new(AuthoringSettingKind.Boolean)),
         ("*", "PublishSummaries", new(AuthoringSettingKind.Boolean)),
-        ("*", "DwellLimitMs", new(AuthoringSettingKind.Integer, Minimum: 0)),
+        ("*", "DwellLimitMs", new(AuthoringSettingKind.Double, Minimum: 0)),
         ("*", "SeriesCompliance", new(AuthoringSettingKind.Choice, AuthoringSettingChoiceSource.SeriesCompliance)),
         ("*", "Channel", new(AuthoringSettingKind.Choice, AuthoringSettingChoiceSource.ChannelKeys, ChoiceIsEditable: true)),
         ("*", "InputChannel", new(AuthoringSettingKind.Choice, AuthoringSettingChoiceSource.ChannelKeys, ChoiceIsEditable: true)),
